third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/select-value-expected.txt

This is a testharness.js-based test.
[FAIL] option is child of div
  assert_equals: expected "1" but got ""
Harness: the test ran to completion.
